move public upload/downlod steps into separate context
