+++ tests/acceptance/config/behat.yml
@@ -74,12 +74,14 @@ default:
         - '%paths.base%/../features/apiShareManagement'
       contexts:
         - FeatureContext: *common_feature_context_params
+        - PublicWebDavContext:
 
     apiShareOperations:
       paths:
         - '%paths.base%/../features/apiShareOperations'
       contexts:
         - FeatureContext: *common_feature_context_params
+        - PublicWebDavContext:
 
     apiSharingNotifications:
       paths:
@@ -107,6 +109,7 @@ default:
         - FeatureContext: *common_feature_context_params
         - LoggingContext:
         - SearchContext:
+        - PublicWebDavContext:
 
     apiWebdavProperties:
       paths:
@@ -202,6 +205,7 @@ default:
         - WebUISearchContext:
         - WebUISharingContext:
         - OccContext:
+        - PublicWebDavContext:
 
     webUILogin:
       paths:
@@ -223,6 +227,7 @@ default:
         - WebUIGeneralContext:
         - WebUILoginContext:
         - WebUISharingContext:
+        - PublicWebDavContext:
 
     webUIPersonalSettings:
       paths:
@@ -247,6 +252,7 @@ default:
         - WebUIGeneralContext:
         - WebUILoginContext:
         - WebUISharingContext:
+        - PublicWebDavContext:
 
     webUIRenameFolders:
       paths:
@@ -278,6 +284,7 @@ default:
         - WebUIGeneralContext:
         - WebUILoginContext:
         - WebUISharingContext:
+        - PublicWebDavContext:
 
     webUISharingInternalGroups:
       paths:
@@ -341,6 +348,7 @@ default:
         - WebUIGeneralContext:
         - WebUILoginContext:
         - WebUISharingContext:
+        - PublicWebDavContext:
 
     # This suite is part of the user_management app in later core versions
     webUIManageQuota:
@@ -390,6 +398,7 @@ default:
         - WebUILoginContext:
         - WebUIWebDavLockingContext:
         - WebUISharingContext:
+        - PublicWebDavContext:
 
   extensions:
     jarnaiz\JUnitFormatter\JUnitFormatterExtension:
