+++ tests/acceptance/features/bootstrap/Sharing.php
@@ -511,74 +511,6 @@ trait Sharing {
 	}
 
 	/**
-	 * @When the public uploads file ":filename" with content ":body" using the public WebDAV API
-	 * @Given the public has uploaded file ":filename" with content ":body"
-	 *
-	 * @param string $filename target file name
-	 * @param string $body content to upload
-	 *
-	 * @return void
-	 */
-	public function publiclyUploadingContent($filename, $body = 'test') {
-		$this->publicUploadContent($filename, '', $body);
-	}
-
-	/**
-	 * @When the public overwrites file ":filename" with content ":body" using the old WebDAV API
-	 * @Given the public has overwritten file ":filename" with content ":body"
-	 *
-	 * @param string $filename target file name
-	 * @param string $body content to upload
-	 *
-	 * @return void
-	 */
-	public function publiclyOverwritingContent($filename, $body = 'test') {
-		$this->publicUploadContent($filename, '', $body, false);
-	}
-
-	/**
-	 * @When the public uploads file ":filename" with password ":password" and content ":body" using the public WebDAV API
-	 * @Given the public has uploaded file ":filename" with password ":password" and content ":body"
-	 *
-	 * @param string $filename target file name
-	 * @param string $password
-	 * @param string $body content to upload
-	 *
-	 * @return void
-	 */
-	public function publiclyUploadingContentWithPassword(
-		$filename, $password = '', $body = 'test'
-	) {
-		$this->publicUploadContent($filename, $password, $body);
-	}
-
-	/**
-	 * @When the public uploads file ":filename" with content ":body" with autorename mode using the public WebDAV API
-	 * @Given the public has uploaded file ":filename" with content ":body" with autorename mode
-	 *
-	 * @param string $filename target file name
-	 * @param string $body content to upload
-	 *
-	 * @return void
-	 */
-	public function publiclyUploadingContentAutorename($filename, $body = 'test') {
-		$this->publicUploadContent($filename, '', $body, true);
-	}
-
-	/**
-	 * @When the public uploads file ":filename" using the old WebDAV API
-	 * @Given the public has uploaded file ":filename"
-	 *
-	 * @param string $source target file name
-	 *
-	 * @return void
-	 */
-	public function publiclyUploadingFile($source) {
-		$file = \GuzzleHttp\Stream\Stream::factory(\fopen($source, 'r'));
-		$this->publicUploadContent(\basename($source), '', $file->getContents());
-	}
-
-	/**
 	 * @Then /^user "([^"]*)" should not be able to create a public link share of (?:file|folder) "([^"]*)" using the sharing API$/
 	 *
 	 * @param string $sharer
@@ -591,69 +523,6 @@ trait Sharing {
 		PHPUnit_Framework_Assert::assertEquals(
 			404,
 			$this->getOCSResponseStatusCode($this->response)
-		);
-	}
-
-	/**
-	 * @Then publicly uploading a file should not work
-	 *
-	 * @return void
-	 */
-	public function publiclyUploadingShouldNotWork() {
-		$this->publicUploadContent('whateverfilefortesting.txt', '', 'test');
-
-		PHPUnit_Framework_Assert::assertTrue(
-			($this->response->getStatusCode() == 507)
-			|| (
-				($this->response->getStatusCode() >= 400)
-				&& ($this->response->getStatusCode() <= 499)
-			),
-			"upload should have failed but passed with code " . $this->response->getStatusCode()
-		);
-	}
-
-	/**
-	 * @Then publicly uploading a file should work
-	 *
-	 * @return void
-	 */
-	public function publiclyUploadingShouldWork() {
-		$path = 'whateverfilefortesting.txt';
-		$content = 'test';
-		$this->publicUploadContent($path, '', $content);
-
-		PHPUnit_Framework_Assert::assertTrue(
-			($this->response->getStatusCode() == 201),
-			"upload should have passed but failed with code " . $this->response->getStatusCode()
-		);
-		$this->shouldBeAbleToDownloadFileInsidePublicSharedFolder("bytes=0-3", $path, $content);
-	}
-
-	/**
-	 * @param string $filename
-	 * @param string $password
-	 * @param string $body
-	 * @param bool $autorename
-	 *
-	 * @return void
-	 */
-	private function publicUploadContent(
-		$filename,
-		$password = '',
-		$body = 'test',
-		$autorename = false
-	) {
-		$password = $this->getActualPassword($password);
-		$url = $this->getBaseUrl() . "/public.php/webdav/";
-		$url .= \rawurlencode(\ltrim($filename, '/'));
-		$token = $this->getLastShareToken();
-		$headers = ['X-Requested-With' => 'XMLHttpRequest'];
-
-		if ($autorename) {
-			$headers['OC-Autorename'] = 1;
-		}
-		$this->response = HttpRequestHelper::put(
-			$url, $token, $password, $headers, $body
 		);
 	}
