+++ tests/acceptance/features/bootstrap/WebDav.php
@@ -536,103 +536,6 @@ trait WebDav {
 	}
 
 	/**
-	 * @When /^the public downloads the last public shared file with range "([^"]*)" using the public WebDAV API$/
-	 *
-	 * @param string $range
-	 *
-	 * @return void
-	 */
-	public function downloadPublicFileWithRange($range) {
-		$token = $this->lastShareData->data->token;
-		$fullUrl = $this->getBaseUrl() . "/public.php/webdav";
-		$headers = [
-			'X-Requested-With' => 'XMLHttpRequest',
-			'Range' => $range
-		];
-		$this->response = HttpRequestHelper::get($fullUrl, $token, "", $headers);
-	}
-
-	/**
-	 * @When /^the public downloads file "([^"]*)" from inside the last public shared folder with range "([^"]*)" using the public WebDAV API$/
-	 *
-	 * @param string $path
-	 * @param string $range
-	 *
-	 * @return void
-	 */
-	public function downloadPublicFileInsideAFolderWithRange($path, $range) {
-		$fullUrl = $this->getBaseUrl() . "/public.php/webdav$path";
-		$headers = [
-			'X-Requested-With' => 'XMLHttpRequest',
-			'Range' => $range
-		];
-		$this->response = HttpRequestHelper::get(
-			$fullUrl, $this->lastShareData->data->token, "", $headers
-		);
-	}
-
-	/**
-	 * @When /^the public downloads file "([^"]*)" from inside the last public shared folder with password "([^"]*)" with range "([^"]*)" using the public WebDAV API$/
-	 *
-	 * @param string $path
-	 * @param string $password
-	 * @param string $range
-	 *
-	 * @return void
-	 */
-	public function publicDownloadsTheFileInsideThePublicSharedFolderWithPassword(
-		$path, $password, $range
-	) {
-		$path = \ltrim($path, "/");
-		$password = $this->getActualPassword($password);
-		$fullUrl = $this->getBaseUrl() . "/public.php/webdav/$path";
-		$headers = [
-			'X-Requested-With' => 'XMLHttpRequest',
-			'Range' => $range
-		];
-		$this->response = HttpRequestHelper::get(
-			$fullUrl, $this->lastShareData->data->token, $password, $headers
-		);
-	}
-	
-	/**
-	 * @Then /^the public should be able to download the range "([^"]*)" of file "([^"]*)" from inside the last public shared folder with password "([^"]*)" and the content should be "([^"]*)"$/
-	 *
-	 * @param string $range
-	 * @param string $path
-	 * @param string $password
-	 * @param string $content
-	 *
-	 * @return void
-	 */
-	public function shouldBeAbleToDownloadFileInsidePublicSharedFolderWithPassword(
-		$range, $path, $password, $content
-	) {
-		$this->publicDownloadsTheFileInsideThePublicSharedFolderWithPassword(
-			$path, $password, $range
-		);
-		$this->downloadedContentShouldBe($content);
-	}
-
-	/**
-	 * @Then /^the public should be able to download the range "([^"]*)" of file "([^"]*)" from inside the last public shared folder and the content should be "([^"]*)"$/
-	 *
-	 * @param string $range
-	 * @param string $path
-	 * @param string $content
-	 *
-	 * @return void
-	 */
-	public function shouldBeAbleToDownloadFileInsidePublicSharedFolder(
-		$range, $path, $content
-	) {
-		$this->publicDownloadsTheFileInsideThePublicSharedFolderWithPassword(
-			$path, null, $range
-		);
-		$this->downloadedContentShouldBe($content);
-	}
-
-	/**
 	 * @Then /^user "([^"]*)" using password "([^"]*)" should not be able to download file "([^"]*)"$/
 	 *
 	 * @param string $user
